feat: initial improvements to make curriculum visualiser flexible
