+++ src/components/CurriculumComponents/CurriculumVisualiserLayout/CurriculumVisualiserLayout.tsx
@@ -1,0 +1,28 @@
+import { OakFlex } from "@oaknational/oak-components";
+import styled from "styled-components";
+
+const CurriculumVisualiserLayoutLeft = styled(OakFlex)`
+  min-width: 296px;
+`;
+
+type CurriculumVisualiserLayoutProps = {
+  filters: React.ReactNode;
+  units: React.ReactNode;
+};
+export default function CurriculumVisualiserLayout({
+  filters,
+  units,
+}: CurriculumVisualiserLayoutProps) {
+  return (
+    <OakFlex>
+      <CurriculumVisualiserLayoutLeft
+        data-test-id="filter-sidebar"
+        $flexDirection={"column"}
+        $display={["none", "block"]}
+      >
+        {filters}
+      </CurriculumVisualiserLayoutLeft>
+      <OakFlex $flexGrow={1}>{units}</OakFlex>
+    </OakFlex>
+  );
+}
