+++ src/components/CurriculumComponents/UnitsTab/UnitsTab.test.tsx
@@ -1,7 +1,7 @@
 import userEvent from "@testing-library/user-event";
 import { waitFor } from "@testing-library/react";
 
-import UnitsTab, { createProgrammeSlug } from "./UnitsTab";
+import UnitsTab from "./UnitsTab";
 
 import { formatCurriculumUnitsData } from "@/pages/teachers/curriculum/[subjectPhaseSlug]/[tab]";
 import curriculumUnitsTabFixture from "@/node-lib/curriculum-api-2023/fixtures/curriculumUnits.fixture";
@@ -857,196 +857,6 @@ describe("components/pages/CurriculumInfo/tabs/UnitsTab", () => {
     );
     const tag = await findByTestId("options-tag");
     expect(tag).toHaveTextContent("2");
-  });
-
-  describe("programme slugs are created correctly", () => {
-    test("unit data ks2 returns correct programme slug", () => {
-      const unitData = {
-        planned_number_of_lessons: 5,
-        connection_future_unit_description: null,
-        connection_prior_unit_description: null,
-        connection_future_unit_title: null,
-        connection_prior_unit_title: null,
-        domain: null,
-        domain_id: null,
-        examboard: null,
-        examboard_slug: null,
-        keystage_slug: "ks2",
-        lessons: [],
-        order: 1,
-        phase: "Primary",
-        phase_slug: "primary",
-        slug: "cellular-respiration-and-atp",
-        subject: "Science",
-        subject_parent: null,
-        subject_parent_slug: null,
-        subject_slug: "science",
-        tags: null,
-        subjectcategories: null,
-        threads: [],
-        tier: null,
-        tier_slug: null,
-        title: "Aerobic and anaerobic cellular respiration",
-        unit_options: [],
-        year: "11",
-        cycle: "1",
-        why_this_why_now: null,
-        description: null,
-        state: "published",
-      };
-      expect(createProgrammeSlug(unitData)).toEqual("science-primary-ks2");
-    });
-    test("unit data ks4 returns correct programme slug", () => {
-      const unitData = {
-        planned_number_of_lessons: 5,
-        connection_future_unit_description: null,
-        connection_prior_unit_description: null,
-        connection_future_unit_title: null,
-        connection_prior_unit_title: null,
-        domain: null,
-        domain_id: null,
-        examboard: null,
-        examboard_slug: null,
-        keystage_slug: "ks4",
-        lessons: [],
-        order: 1,
-        phase: "Primary",
-        phase_slug: "primary",
-        slug: "cellular-respiration-and-atp",
-        subject: "Science",
-        subject_parent: null,
-        subject_parent_slug: null,
-        subject_slug: "science",
-        tags: null,
-        subjectcategories: null,
-        threads: [],
-        tier: null,
-        tier_slug: null,
-        title: "Aerobic and anaerobic cellular respiration",
-        unit_options: [],
-        year: "11",
-        cycle: "1",
-        why_this_why_now: null,
-        description: null,
-        state: "published",
-      };
-      expect(createProgrammeSlug(unitData)).toEqual("science-primary-ks4");
-    });
-    test("unit data with exam board and tier returns the correct programme slug", () => {
-      const unitData = {
-        planned_number_of_lessons: 5,
-        connection_future_unit_description: null,
-        connection_prior_unit_description: null,
-        connection_future_unit_title: null,
-        connection_prior_unit_title: null,
-        domain: null,
-        domain_id: null,
-        examboard: null,
-        examboard_slug: null,
-        keystage_slug: "ks4",
-        lessons: [],
-        order: 1,
-        phase: "Secondary",
-        phase_slug: "secondary",
-        slug: "cellular-respiration-and-atp",
-        subject: "Combined Science",
-        subject_parent: "Science",
-        subject_parent_slug: "science",
-        subject_slug: "combined-science",
-        tags: null,
-        subjectcategories: null,
-        threads: [],
-        tier: null,
-        tier_slug: null,
-        title: "Aerobic and anaerobic cellular respiration",
-        unit_options: [],
-        year: "11",
-        cycle: "1",
-        why_this_why_now: null,
-        description: null,
-        state: "published",
-      };
-      expect(createProgrammeSlug(unitData, "aqa", "foundation")).toEqual(
-        "combined-science-secondary-ks4-foundation-aqa",
-      );
-    });
-    test("unit data for ks3 returns the correct programme slug", () => {
-      const unitData = {
-        planned_number_of_lessons: 5,
-        connection_future_unit_description: null,
-        connection_prior_unit_description: null,
-        connection_future_unit_title: null,
-        connection_prior_unit_title: null,
-        domain: null,
-        domain_id: null,
-        examboard: null,
-        examboard_slug: null,
-        keystage_slug: "ks3",
-        lessons: [],
-        order: 1,
-        phase: "Secondary",
-        phase_slug: "secondary",
-        slug: "cellular-respiration-and-atp",
-        subject: "Combined Science",
-        subject_parent: "Science",
-        subject_parent_slug: "science",
-        subject_slug: "combined-science",
-        tags: null,
-        subjectcategories: null,
-        threads: [],
-        tier: null,
-        tier_slug: null,
-        title: "Aerobic and anaerobic cellular respiration",
-        unit_options: [],
-        year: "9",
-        cycle: "1",
-        why_this_why_now: null,
-        description: null,
-        state: "published",
-      };
-      expect(createProgrammeSlug(unitData, "aqa")).toEqual(
-        "combined-science-secondary-ks3",
-      );
-    });
-
-    test("unit data for ks3 returns the correct programme slug", () => {
-      const unitData = {
-        planned_number_of_lessons: 5,
-        connection_future_unit_description: null,
-        connection_prior_unit_description: null,
-        connection_future_unit_title: null,
-        connection_prior_unit_title: null,
-        domain: null,
-        domain_id: null,
-        examboard: null,
-        examboard_slug: null,
-        keystage_slug: "ks3",
-        lessons: [],
-        order: 1,
-        phase: "Secondary",
-        phase_slug: "secondary",
-        slug: "cellular-respiration-and-atp",
-        subject: "Combined Science",
-        subject_parent: "Science",
-        subject_parent_slug: "science",
-        subject_slug: "combined-science",
-        tags: null,
-        subjectcategories: null,
-        threads: [],
-        tier: null,
-        tier_slug: null,
-        title: "Aerobic and anaerobic cellular respiration",
-        unit_options: [],
-        year: "9",
-        cycle: "1",
-        why_this_why_now: null,
-        description: null,
-        state: "published",
-      };
-      expect(createProgrammeSlug(unitData, "aqa")).toEqual(
-        "combined-science-secondary-ks3",
-      );
-    });
   });
 
   const resizeWindow = (x: number, y: number) => {
